% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{Change}
\alias{Change}
\title{Google Cloud DNS API Objects 
Configures and serves authoritative DNS records.}
\usage{
Change(additions = NULL, deletions = NULL, id = NULL, isServing = NULL,
  startTime = NULL, status = NULL)
}
\arguments{
\item{additions}{Which ResourceRecordSets to add?}

\item{deletions}{Which ResourceRecordSets to remove? Must match existing data exactly}

\item{id}{Unique identifier for the resource; defined by the server (output only)}

\item{isServing}{If the DNS queries for the zone will be served}

\item{startTime}{The time that this operation was started by the server (output only)}

\item{status}{Status of the operation (output only)}
}
\value{
Change object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:11:00
filename: /Users/mark/dev/R/autoGoogleAPI/googlednsv2beta1.auto/R/dns_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Change Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An atomic update to a collection of ResourceRecordSets.
}
\seealso{
Other Change functions: \code{\link{changes.create}}
}

