% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_functions.R
\name{changes.get}
\alias{changes.get}
\title{Fetch the representation of an existing Change.}
\usage{
changes.get(project, managedZone, changeId, clientOperationId = NULL)
}
\arguments{
\item{project}{Identifies the project addressed by this request}

\item{managedZone}{Identifies the managed zone addressed by this request}

\item{changeId}{The identifier of the requested change, from a previous ResourceRecordSetsChangeResponse}

\item{clientOperationId}{For mutating operation requests only}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/ndev.clouddns.readonly
\item https://www.googleapis.com/auth/ndev.clouddns.readwrite
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only, https://www.googleapis.com/auth/ndev.clouddns.readonly, https://www.googleapis.com/auth/ndev.clouddns.readwrite)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/cloud-dns}{Google Documentation}
}

