% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_functions.R
\docType{package}
\name{dns_googleAuthR}
\alias{dns_googleAuthR}
\alias{dns_googleAuthR-package}
\title{Google Cloud DNS API
Configures and serves authoritative DNS records.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:48:17
filename: /Users/mark/dev/R/autoGoogleAPI/googlednsv2beta1.auto/R/dns_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/ndev.clouddns.readonly
\item https://www.googleapis.com/auth/ndev.clouddns.readwrite
}
}
