% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{DnsKey}
\alias{DnsKey}
\title{DnsKey Object}
\usage{
DnsKey(algorithm = NULL, creationTime = NULL, description = NULL,
  digests = NULL, id = NULL, isActive = NULL, keyLength = NULL,
  keyTag = NULL, publicKey = NULL, type = NULL)
}
\arguments{
\item{algorithm}{String mnemonic specifying the DNSSEC algorithm of this key}

\item{creationTime}{The time that this resource was created in the control plane}

\item{description}{A mutable string of at most 1024 characters associated with this resource for the user's convenience}

\item{digests}{Cryptographic hashes of the DNSKEY resource record associated with this DnsKey}

\item{id}{Unique identifier for the resource; defined by the server (output only)}

\item{isActive}{Active keys will be used to sign subsequent changes to the ManagedZone}

\item{keyLength}{Length of the key in bits}

\item{keyTag}{The key tag is a non-cryptographic hash of the a DNSKEY resource record associated with this DnsKey}

\item{publicKey}{Base64 encoded public half of this key}

\item{type}{One of 'KEY_SIGNING' or 'ZONE_SIGNING'}
}
\value{
DnsKey object
}
\description{
DnsKey Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A DNSSEC key pair.
}
