% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{DnsKeyDigest}
\alias{DnsKeyDigest}
\title{DnsKeyDigest Object}
\usage{
DnsKeyDigest(digest = NULL, type = NULL)
}
\arguments{
\item{digest}{The base-16 encoded bytes of this digest}

\item{type}{Specifies the algorithm used to calculate this digest}
}
\value{
DnsKeyDigest object
}
\description{
DnsKeyDigest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
