% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{DnsKeysListResponse}
\alias{DnsKeysListResponse}
\title{DnsKeysListResponse Object}
\usage{
DnsKeysListResponse(dnsKeys = NULL, header = NULL, nextPageToken = NULL)
}
\arguments{
\item{dnsKeys}{The requested resources}

\item{header}{No description}

\item{nextPageToken}{The presence of this field indicates that there exist more results following your last page of results in pagination order}
}
\value{
DnsKeysListResponse object
}
\description{
DnsKeysListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response to a request to enumerate DnsKeys in a ManagedZone.
}
