% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{DnsKeySpec}
\alias{DnsKeySpec}
\title{DnsKeySpec Object}
\usage{
DnsKeySpec(algorithm = NULL, keyLength = NULL, keyType = NULL)
}
\arguments{
\item{algorithm}{String mnemonic specifying the DNSSEC algorithm of this key}

\item{keyLength}{Length of the keys in bits}

\item{keyType}{One of 'KEY_SIGNING' or 'ZONE_SIGNING'}
}
\value{
DnsKeySpec object
}
\description{
DnsKeySpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parameters for DnsKey key generation. Used for generating initial keys for a new ManagedZone and as default when adding a new DnsKey.
}
