% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{ManagedZoneDnsSecConfig}
\alias{ManagedZoneDnsSecConfig}
\title{ManagedZoneDnsSecConfig Object}
\usage{
ManagedZoneDnsSecConfig(defaultKeySpecs = NULL, nonExistence = NULL,
  state = NULL)
}
\arguments{
\item{defaultKeySpecs}{Specifies parameters that will be used for generating initial DnsKeys for this ManagedZone}

\item{nonExistence}{Specifies the mechanism used to provide authenticated denial-of-existence responses}

\item{state}{Specifies whether DNSSEC is enabled, and what mode it is in}
}
\value{
ManagedZoneDnsSecConfig object
}
\description{
ManagedZoneDnsSecConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

