% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{ManagedZoneOperationsListResponse}
\alias{ManagedZoneOperationsListResponse}
\title{ManagedZoneOperationsListResponse Object}
\usage{
ManagedZoneOperationsListResponse(header = NULL, nextPageToken = NULL,
  operations = NULL)
}
\arguments{
\item{header}{No description}

\item{nextPageToken}{The presence of this field indicates that there exist more results following your last page of results in pagination order}

\item{operations}{The operation resources}
}
\value{
ManagedZoneOperationsListResponse object
}
\description{
ManagedZoneOperationsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
