% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_functions.R
\name{managedZones.list}
\alias{managedZones.list}
\title{Enumerate ManagedZones that have been created but not yet deleted.}
\usage{
managedZones.list(project, dnsName = NULL, maxResults = NULL,
  pageToken = NULL)
}
\arguments{
\item{project}{Identifies the project addressed by this request}

\item{dnsName}{Restricts the list to return only zones with this domain name}

\item{maxResults}{Optional}

\item{pageToken}{Optional}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/ndev.clouddns.readonly
\item https://www.googleapis.com/auth/ndev.clouddns.readwrite
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only, https://www.googleapis.com/auth/ndev.clouddns.readonly, https://www.googleapis.com/auth/ndev.clouddns.readwrite)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/cloud-dns}{Google Documentation}
}

