% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{ManagedZonesDeleteResponse}
\alias{ManagedZonesDeleteResponse}
\title{ManagedZonesDeleteResponse Object}
\usage{
ManagedZonesDeleteResponse(header = NULL)
}
\arguments{
\item{header}{No description}
}
\value{
ManagedZonesDeleteResponse object
}
\description{
ManagedZonesDeleteResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
