% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{ManagedZonesListResponse}
\alias{ManagedZonesListResponse}
\title{ManagedZonesListResponse Object}
\usage{
ManagedZonesListResponse(header = NULL, managedZones = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{header}{No description}

\item{managedZones}{The managed zone resources}

\item{nextPageToken}{The presence of this field indicates that there exist more results following your last page of results in pagination order}
}
\value{
ManagedZonesListResponse object
}
\description{
ManagedZonesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
