% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{Operation}
\alias{Operation}
\title{Operation Object}
\usage{
Operation(dnsKeyContext = NULL, id = NULL, startTime = NULL,
  status = NULL, type = NULL, user = NULL, zoneContext = NULL)
}
\arguments{
\item{dnsKeyContext}{Only populated if the operation targeted a DnsKey (output only)}

\item{id}{Unique identifier for the resource}

\item{startTime}{The time that this operation was started by the server}

\item{status}{Status of the operation}

\item{type}{Type of the operation}

\item{user}{User who requested the operation, for example: user@example}

\item{zoneContext}{Only populated if the operation targeted a ManagedZone (output only)}
}
\value{
Operation object
}
\description{
Operation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An operation represents a successful mutation performed on a Cloud DNS resource. Operations provide: - An audit log of server resource mutations. - A way to recover/retry API calls in the case where the response is never received by the caller. Use the caller specified client_operation_id.
}

