% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{OperationDnsKeyContext}
\alias{OperationDnsKeyContext}
\title{OperationDnsKeyContext Object}
\usage{
OperationDnsKeyContext(newValue = NULL, oldValue = NULL)
}
\arguments{
\item{newValue}{The post-operation DnsKey resource}

\item{oldValue}{The pre-operation DnsKey resource}
}
\value{
OperationDnsKeyContext object
}
\description{
OperationDnsKeyContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

