% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{OperationManagedZoneContext}
\alias{OperationManagedZoneContext}
\title{OperationManagedZoneContext Object}
\usage{
OperationManagedZoneContext(newValue = NULL, oldValue = NULL)
}
\arguments{
\item{newValue}{The post-operation ManagedZone resource}

\item{oldValue}{The pre-operation ManagedZone resource}
}
\value{
OperationManagedZoneContext object
}
\description{
OperationManagedZoneContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

