% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{ResourceRecordSetsListResponse}
\alias{ResourceRecordSetsListResponse}
\title{ResourceRecordSetsListResponse Object}
\usage{
ResourceRecordSetsListResponse(header = NULL, nextPageToken = NULL,
  rrsets = NULL)
}
\arguments{
\item{header}{No description}

\item{nextPageToken}{The presence of this field indicates that there exist more results following your last page of results in pagination order}

\item{rrsets}{The resource record set resources}
}
\value{
ResourceRecordSetsListResponse object
}
\description{
ResourceRecordSetsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
