% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_objects.R
\name{ResponseHeader}
\alias{ResponseHeader}
\title{ResponseHeader Object}
\usage{
ResponseHeader(operationId = NULL)
}
\arguments{
\item{operationId}{For mutating operation requests that completed successfully}
}
\value{
ResponseHeader object
}
\description{
ResponseHeader Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Elements common to every response.
}

